dump,restore: Use rt_sigaction_t for sys_sigaction
Since we operate with syscalls directly we are
to convert signal's structures between image and
kernel formats, without intermediate glibc layer.

Note this involves chaging sa_entry::flags to u64
(since it's long int value in kernel).

+++ include/image.h
@@ -90,7 +90,7 @@ struct page_entry {
 struct sa_entry {
 	u64		sigaction;
 	u8		mask[128];
-	u32		flags;
+	u64		flags;
 	u64		restorer;
 } __packed;
